+++ tests/test_fields.py
@@ -1,6 +1,5 @@
 from mex.common.fields import (
     ALL_MODEL_CLASSES_BY_NAME,
-    EMAIL_FIELDS_BY_CLASS_NAME,
     FINAL_FIELDS_BY_CLASS_NAME,
     FROZEN_FIELDS_BY_CLASS_NAME,
     INTEGER_FIELDS_BY_CLASS_NAME,
@@ -20,7 +19,6 @@ def test_all_fields_by_class_names_include_all_classes() -> None:
     assert {
         len(lookup)
         for lookup in (
-            EMAIL_FIELDS_BY_CLASS_NAME,
             FINAL_FIELDS_BY_CLASS_NAME,
             FROZEN_FIELDS_BY_CLASS_NAME,
             INTEGER_FIELDS_BY_CLASS_NAME,